Finishing touch

+++ routes/web.php
@@ -13,8 +13,7 @@
 
 Route::get('/', 'HomePageController@home')->name('home');
 Route::get('gallery', 'HomePageController@gallery')->name('gallery');
-Route::get('places', 'HomePageController@places')->name('places');
-Route::get('booking', 'HomePageController@booking')->name('booking');
+Route::get('booking/{name}', 'HomePageController@booking')->name('booking');
 Route::post('booking', 'HomePageController@storebooking');
 Route::get('about', 'HomePageController@about')->name('about');
 Route::get('contact', 'HomePageController@createContact')->name('contact');
@@ -24,12 +23,12 @@ Route::get('package/{name}', 'HomePageController@showCategory')->name('show.cate
 
 Route::get('admin', 'AdminController@getLoginForm')->name('login');
 Route::post('login', 'AdminController@login');
-Route::get('logout', 'AdminController@logout');
-Route::get('messages','AdminController@getMessages');
-Route::delete('message-delete/{id}', 'AdminController@deleteMessage')->name('contact.destroy');
-Route::get('bookings','AdminController@getBookings');
+Route::get('logout', 'AdminController@logout')->middleware('admin');
+Route::get('messages','AdminController@getMessages')->middleware('admin');
+Route::delete('message-delete/{id}', 'AdminController@deleteMessage')->name('contact.destroy')->middleware('admin');
+Route::get('bookings','AdminController@getBookings')->middleware('admin');
 Route::delete('booking-delete/{id}', 'AdminController@deleteBooking')->name('booking.destroy');
 Route::resource('posts', 'PostController')->except(['edit'])->middleware('admin');
-Route::get('posts/{slug}/edit','PostController@edit')->name('posts.edit');
+Route::get('posts/{slug}/edit','PostController@edit')->name('posts.edit')->middleware('admin');
 Route::resource('categories', 'CategoryController')->middleware('admin');
